backend/requirements.txt
# MediVan AI — Python Dependencies
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
httpx>=0.25.0

# ML / Inference
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
open-clip-torch>=2.24.0

# RAG
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
